Cards now start in the center
Previous check in was incorrect. Cards still COULD fall through the table. It had to do with the way I was teleporting cards in Card.cpp. The correct way to do this is TeleportTo, not SetActorLocation (which apparently turns off collision or something). Additionally, all cards now appear at the center of the table. Next step is to try and deal them out!