Added a second Static Mesh to card blueprint. Added material to each of its children classes. I now have cards with a back and a front.
Additionally got to a somewhat stable state. For some reason, if UE5 crashes, it forgets how everything compiles in the Delaware2GameState class. I have to comment out everything that involves the struct and the enum just above it, then slowly un-comment everything until I'm back to square one. Annoying, but I'm hoping as I move further away from modifying it, it'll go away. It's one of the reasons I went to working on the card stuff.