Cards are now given proper IDs to reference later.
Not sure if this was referenced in the last commit, but I added more Targets for cards (before there was just deal to, now there is deck and deal from).

Moved the FDealLocationVector struct into its own header. Hoping that the random, ridiculous crashes will stop happening with it (and its associated enum) on its own header.