Added a new static mesh for cards that puts both card back and card face on the same mesh.
Cards deal!

Added a new static function for the FDealLocationVectors struct:
GetDealingLocationName which is essentially a ->ToString() function

Added a new const static array of strings EPlayerAsString, which is essentially a ->ToString() function

Fixed a bug in EPlayers ++ operator as it was not resetting back to North after West like it should.