I did two things to try and fix the cards not all sorting into place as they should: 1) I tried to prevent the cards from 'sticking' into the table. I think this was caused when they had to great of a pitch/roll. I prevented them from getting too far over from one or the other. 2) I added an additional sort command at the end of dealing, just in case there is some miscellaneous card still in the center.
Also included in this fix is that I moved all of the blocking volumes further away from the table. I had to do this because when South was dealing to himself, the cards would often get pitched and looked wonky. I did this with the other blocking volumes as well, but since I don't have a way of determining if it's happening to them, I was unable to test them. When I moved the blocking volumes, I also had to add another blocking volume, which is more like an extension of the table so cards to do not fall through.