Bid Bubbles are showing!
Had to move each of the phases to be part of the TableTopActor class. It was hard to do with GameState because GameState only appears when the game starts. The Dealing and Bidding are kind of part of the table anyway. With this, I was able to place the UI into the table within the editor, making the code work. Next up is to make the bubbles appear when I want them to